Add files via upload
Crypto pair correlator. Choose any two coins that are listed on CoinGecko and plot the prices, price ratio and correlation over a variety of periods. Inputs are specified in the "PairCorrInput.csv" file. The coin-gecko API ID name must be used in the A2 and B2 cells to specify the coins.  Enter the number of days for which you want to calculate the data in cell C2. In column D starting from D2, enter numbers to indicate the periods over which you wish to calculate the correlation moving averages. This can be arbitray numbers and number of averages yet having too many will make the correlation graph to busy to be useful. An output file is generated that contains price data and the correlation and covariance moving averages. You will of course need the python packages that are imported by the file at the top of the script, already installed on your python distribution. You'll also need to specify the directory on your computer containing the input and output files where necessary in the code (currently lines 15 and 144). 
  
This tool should be useful for assessing whether a given pair of coins are a good match for yield farming or whether you'll get smashed by impermanent loss. I'm aiming to add a panel interface and make into a self-contained application eventually.  This was tested only on mac with python3, thus far.